refactor: width는 string형식으로 받도록 수정 및 파일 이름 변경
InputWrapper -> InputContainer로 파일이름 변경
